+++ packages/ricos-content/web/src/TranslatableAPI/types.ts
@@ -1,5 +1,4 @@
 import type { PropertyPath } from 'lodash';
-import { Optional } from 'monocle-ts';
 import type { Node } from 'ricos-schema';
 import { Node_Type } from 'ricos-schema';
 import type { RICOS_NODE_TYPE_TO_DATA_FIELD } from '../consts';