+++ packages/ricos-content/web/src/velo-adapter/query.ts
@@ -2,11 +2,7 @@ import { get } from 'lodash';
 import type { RichContent, Node } from 'ricos-schema';
 import type { Extractor } from '../extractor-infra';
 import { extract } from '../RicosContentAPI/extract';
-
-const strCompare = (str1: string, str2: string) => str1.localeCompare(str2);
-const numCompare = (num1: number, num2: number) => num1 - num2;
-const compare = (input: string | number, current) =>
-  typeof input === 'string' ? strCompare(current, input) : numCompare(current, input);
+import { compare } from './utils';
 
 class Query {
   private extractor: Extractor<Node>;